# socialbu_api_docs.yaml
openapi: 3.1.0
info:
    description: |
        This is documentation for SocialBu API
        [http://socialbu.com](http://socialbu.com).
    version: "1.0.0"
    title: SocialBu API
    termsOfService: 'https://socialbu.com/about/terms'
    contact:
        email: support@socialbu.com
servers:
    - description: SocialBu API
      url: https://socialbu.com/api/v1
...
security:
    - bearerAuth: []
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
...
paths:
...
    '/posts/{postId}':
        get:
            tags:
                - Posts
            summary: Find post by ID
            description: Returns a single post
            operationId: getPost
            parameters:
                - name: postId
                  in: path
                  description: ID of post to return
                  required: true
                  schema:
                      type: integer
            responses:
                '400':
                    description: Bad Request
                '404':
                    description: Resource not found
                '405':
                    description: Validation exception
                '200':
                    description: Post found
                    content:
                        application/json:
                            schema:
                                type: object
                                description: An object containing the details of the post.
        delete:
            tags:
                - Posts
            summary: Deletes the post
            operationId: deletePost
            parameters:
                - name: postId
                  in: path
                  description: Post id to delete
                  required: true
                  schema:
                      type: integer
            responses:
                '400':
                    description: Bad Request
                '404':
                    description: Resource not found
                '405':
                    description: Validation exception
                '200':
                    description: Post deleted
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        description: Indicates whether the post was deleted.
                                    message:
                                        type: string
                                        description: A message confirming that the post was deleted successfully.
        patch:
            tags:
                - Posts
            summary: Update the post
            operationId: updatePost
            parameters:
                - name: postId
                  in: path
                  description: Post id to update
                  required: true
                  schema:
                      type: integer
            responses:
                '404':
                    description: Resource not found
                '422':
                    description: The given data was invalid.
                '200':
                    description: Post updated
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        description: Indicates whether the post was updated successfully.
                                    message:
                                        type: string
                                        description: A message confirming that the post was updated successfully.
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                accounts:
                                    type: array
                                    items:
                                        type: integer
                                    description: ID(s) of one or more your social accounts.
                                team_id:
                                    type: integer
                                    description: Unique ID of your team.
                                publish_at:
                                    example: "2025-04-14 15:30:00"
                                    type: string
                                    description: Format should be Y-m-d H:i:s (UTC)
                                content:
                                    type: string
                                    description: Updated content of the post.
                                existing_attachments:
                                    type: array
                                    items:
                                        type: string
                                        description: The value received from /upload_media/status endpoint.
                                options:
                                    type: object
                                    description: Additional options for the post.
                                    properties:
                                        video_title:
                                            type: string
                                            description: Title of the video
                                        trim_link_from_content:
                                            type: boolean
                                            description: Trim the link from the content.
                                        link:
                                            type: string
                                            description: Link of your media.
                                        media_alt_text:
                                            type: string
                                            description: Alternate text descriptions for media.
                                            items:
                                                type: string
                                        comment:
                                            type: string
                                            description: First comment for the post.
...